Dodata autentifikacija Usera

+++ app/Http/Controllers/PorudzbineController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\ShopingList;
 use App\Models\Post;
 use App\Http\Requests\UpadateUser;
+use Exception;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Auth;
 
@@ -15,19 +16,34 @@ class PorudzbineController extends Controller
     {
         return view('porudzbine');
     }
-    public function getPorudzbine()
+    public function getPorudzbine(Request $request)
     {
-        return ShopingList::all();
+        $user = $request->user();
+        return ShopingList::where('user_id', $user->id)->get();
     }
-    public function prikaziPorudzbinu($id)
+    public function prikaziPorudzbinu($id, Request $request)
     {
-        $porudzbina=ShopingList::findOrFail($id);
+        $user = $request->user();
+        
+        $porudzbina = ShopingList::findOrFail($id);
+
+        if ($porudzbina->user_id != $user->id) {
+            return redirect()->route('viewPorudzbine')->with('danger', 'Ne mozes da vidis tudj proizvod'); 
+        }
+
         return view('porudzbina',compact('porudzbina'));
     }
 
-    public function deletePorudzbine($id)
+    public function deletePorudzbine($id, Request $request)
     {
+        $user = $request->user();
+
         $porudzbina = ShopingList::findOrFail($id);
+
+        if ($porudzbina->user_id != $user->id) 
+        {
+            throw new Exception('Nemas pravo da obrises tudj proizvod');
+        }
 
         $porudzbina->delete();
         return 'ok';
@@ -45,10 +61,12 @@ class PorudzbineController extends Controller
             'cena'=>'required',
         ]);
 
+        $user = $request->user();
+
         $proizvod = new ShopingList;
         $proizvod->ime = $data['ime'];
         $proizvod->cena = $data['cena'];
-        $proizvod->user_id = 1;
+        $proizvod->user_id = $user->id;
 
         $proizvod->save();
         return $proizvod;
@@ -62,7 +80,13 @@ class PorudzbineController extends Controller
 
     public function updatePorudzbine($id, Request $request)
     {
+        $user = $request->user();
         $porudzbina = ShopingList::findOrFail($id);
+
+        if ($porudzbina->user_id != $user->id) {
+            // TODO: Greska
+            throw new Exception('Ne mozes da menjas tudji proizvod');
+        }
 
         $data = $request->validate([
             'ime'=>'required',
@@ -77,8 +101,9 @@ class PorudzbineController extends Controller
     
     public function pretragaPorudzbina(Request $request)
     {
+        $user = $request->user();
         $upit = $request->input('upit', '');
-        $proizvodi = ShopingList::where('ime', 'like', '%' . $upit . '%')->get();
+        $proizvodi = ShopingList::where('user_id', $user->id)->where('ime', 'like', '%' . $upit . '%')->get();
         return $proizvodi;
     }
 }
